Refactor nft validation architecture folder
