+++ client/src/pages/NftValidationPage/components/ContractAddressTitle.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import GenericTitleComponent from './GenericTitleComponent';
 
-import { TITLE_NAME } from '../../constant';
-import classes from './ContractAddressTitle.module.scss';
-import { RenderTitle } from '../types';
+import { TITLE_NAME } from '../constant';
+import classes from '../styles/ContractAddressTitle.module.scss';
+import { RenderTitle } from './types';
 
 type Props = {
   values: { chainCode?: string; contractAddress?: string; tokenId?: string };
