+++ client/src/pages/NftValidationPage/constant.tsx
@@ -4,19 +4,19 @@ import ContractAddressField from './components/ContractAddressField';
 import GenericNftValidationField, {
   FIELDS_NAMES,
 } from './components/GenericNftValidationField';
-import { renderContractAddressTitle } from './components/NftListTitle/ContractAddressTitle';
+import { renderContractAddressTitle } from './components/ContractAddressTitle';
 import {
   renderFioAddressTitle,
   renderHashTitle,
   renderImageTitle,
-} from './components/NftListTitle/GenericTitleComponent';
+} from './components/GenericTitleComponent';
 
 import {
   renderContractAddressItem,
   renderFioAddressItem,
   renderHashItem,
   renderImageItem,
-} from './components/NftItemResult/GenericNftItemResult';
+} from './components/GenericNftItemResult';
 
 import ImageField from './components/ImageField';
