+++ client/src/pages/NftValidationPage/styles/ContractAddressTitle.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../assets/styles/mediaQueriesMixings.scss';
+@import '../../../assets/styles/mediaQueriesMixings.scss';
 
 .container {
   display: flex;
