+++ client/src/pages/NftValidationPage/styles/GenericNftValidationField.module.scss
