+++ client/src/pages/NftValidationPage/styles/GenericTitleComponent.module.scss
@@ -1,6 +1,6 @@
-@import '../../../../assets/styles/colors.scss';
-@import '../../../../assets/styles/fontMixins.scss';
-@import '../../../../assets/styles/mediaQueriesMixings.scss';
+@import '../../../assets/styles/colors.scss';
+@import '../../../assets/styles/fontMixins.scss';
+@import '../../../assets/styles/mediaQueriesMixings.scss';
 
 .container {
   display: flex;
