+++ client/src/pages/NftValidationPage/styles/ImageTitle.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../assets/styles/mediaQueriesMixings.scss';
+@import '../../../assets/styles/mediaQueriesMixings.scss';
 
 .image {
   margin-bottom: 0;
