+++ client/src/pages/NftValidationPage/styles/NftListResults.module.scss
