+++ client/src/pages/NftValidationPage/styles/NftValidationForm.module.scss
