+++ client/src/pages/NftValidationPage/styles/NftValidationItemModal.module.scss
